Fix evaluator

+++ code/src/transformer/evaluator.py
@@ -154,37 +154,38 @@ def main():
     
 
     while True:
-        unevaluated = [f for f in os.listdir(MODEL_CHECKPOINT_PATH) if f.startswith('ckpt')]
-
-        for unevaluated_checkpoint in unevaluated:
-            step = int(unevaluated_checkpoint[5:])
-            result_dir = os.path.join(MODEL_CHECKPOINT_PATH, "results")
-
-            model.load_weights(os.path.join(MODEL_CHECKPOINT_PATH, unevaluated_checkpoint)).expect_partial()
+        if os.path.isdir(MODEL_CHECKPOINT_PATH):
+            unevaluated = [f for f in os.listdir(MODEL_CHECKPOINT_PATH) if f.startswith('ckpt')]
     
-            predicted_sentences = []
-
-            for i, batch in enumerate(dataset):
-                print(f"Evaluating {i+1}. batch...") 
-                preds = model.generate(batch['input_ids'])
-                batch_sentences = tokenizer.batch_decode(preds, skip_special_tokens=True)
-                predicted_sentences = predicted_sentences + batch_sentences
-
-            p, r, f1 = batch_multi_pre_rec_f1(predicted_sentences, dev_source_sentences, dev_gold_edits, 
-                                              MAX_UNCHANGED_WORDS, BETA, IGNORE_WHITESPACE_CASING, VERBOSE, VERY_VERBOSE)
-            
-            file_writer = tf.summary.create_file_writer(result_dir)
-            with file_writer.as_default():
-                tf.summary.scalar('epoch_precision', p, step)
-                tf.summary.scalar('epoch_recall', r, step)
-                tf.summary.scalar('epoch_f1', f1, step)
-
-                text = "\n".join(predicted_sentences[0:20])
-                print(text)
-                tf.summary.text("predictions", text, step)
-
-
-            shutil.rmtree(os.path.join(MODEL_CHECKPOINT_PATH, unevaluated_checkpoint))
+            for unevaluated_checkpoint in unevaluated:
+                step = int(unevaluated_checkpoint[5:])
+                result_dir = os.path.join(MODEL_CHECKPOINT_PATH, "results")
+    
+                model.load_weights(os.path.join(MODEL_CHECKPOINT_PATH, unevaluated_checkpoint)).expect_partial()
+        
+                predicted_sentences = []
+    
+                for i, batch in enumerate(dataset):
+                    print(f"Evaluating {i+1}. batch...") 
+                    preds = model.generate(batch['input_ids'])
+                    batch_sentences = tokenizer.batch_decode(preds, skip_special_tokens=True)
+                    predicted_sentences = predicted_sentences + batch_sentences
+    
+                p, r, f1 = batch_multi_pre_rec_f1(predicted_sentences, dev_source_sentences, dev_gold_edits, 
+                                                  MAX_UNCHANGED_WORDS, BETA, IGNORE_WHITESPACE_CASING, VERBOSE, VERY_VERBOSE)
+                
+                file_writer = tf.summary.create_file_writer(result_dir)
+                with file_writer.as_default():
+                    tf.summary.scalar('epoch_precision', p, step)
+                    tf.summary.scalar('epoch_recall', r, step)
+                    tf.summary.scalar('epoch_f1', f1, step)
+    
+                    text = "\n".join(predicted_sentences[0:20])
+                    print(text)
+                    tf.summary.text("predictions", text, step)
+    
+    
+                shutil.rmtree(os.path.join(MODEL_CHECKPOINT_PATH, unevaluated_checkpoint))
 
         time.sleep(10)
